[P2P-199] fix: 코스(장소) 상세 - 댓글 간 갯수 상이한 점 수정

+++ src/main/java/com/prgrms/p2p/domain/place/dto/DetailPlaceResponse.java
@@ -57,5 +57,5 @@ public class DetailPlaceResponse {
   private String kakaoMapId;
 
   @ApiParam(value = "댓글 개수", example = "0")
-  private Integer comments;
+  private Long comments;
 }
